Wrapped SqlExceptionMapper in SqlExceptionConverter
